refactor: extract stratagem page logic
